+++ src/promptflow/CHANGELOG.md
@@ -10,7 +10,7 @@
 - [SDK/CLI][azure] Create a run with `resume_from`.
   - CLI: Support `pfazure run create --resume-from <original-run-name>` to create a run resume from another run.
   - SDK: Support `pf.run(resume_from=<original-run-name>)` to create a run resume from another run.
-- [SDK] Support flow exectue as async function: `load_flow(<flow>, async_call=True)` will return an async callable flow object.
+- [SDK] Support flow exectue as async function: `load_flow(<flow>, is_async_call=True)` will return an async callable flow object.
 
 ## 1.6.0 (2024.03.01)